+++ kwimage/__init__.py
@@ -67,7 +67,7 @@ __ignore__ = [
     'TORCH_GRID_SAMPLE_HAS_ALIGN',
 ]
 
-__version__ = '0.11.7'
+__version__ = '0.12.0'
 
 import lazy_loader